+++ packages/eslint-plugin/src/Rules/Support/Helpers.ts
@@ -30,10 +30,12 @@ export function meta({
   description,
   schema = [],
   ruleName,
+  type,
 }: RuleMetaDefinition): Pick<Rule.RuleModule, "meta"> {
   return {
     meta: {
-      type: "problem",
+      // TODO: Consider changing default to "suggestion" in a future major release, because most rules are not fixing code issues, but only suggesting best practices.
+      type: type || "problem",
       docs: {
         url: docsUrl(ruleName),
         description,